+++ src/export/schema.cpp
@@ -266,7 +266,7 @@ extern "C" PLUGIN_API void SetEntDataEnt2(CEntityInstance* entity, int offset, i
 		SafeNetworkStateChanged(reinterpret_cast<intptr_t>(entity), offset, chainOffset);
 	}
 
-	*reinterpret_cast<CEntityHandle*>(reinterpret_cast<intptr_t>(entity) + offset) = CEntityHandle((uint32) value);
+	*reinterpret_cast<CEntityHandle*>(reinterpret_cast<intptr_t>(entity) + offset) = CEntityHandle(value);
 }
 
 /**
@@ -292,7 +292,7 @@ extern "C" PLUGIN_API void ChangeEntityState2(CEntityInstance* entity, int offse
  * @return The integer value at the given memory location.
  */
 extern "C" PLUGIN_API int64_t GetEntData(int entityHandle, int offset, int size) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot get '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return {};
@@ -312,7 +312,7 @@ extern "C" PLUGIN_API int64_t GetEntData(int entityHandle, int offset, int size)
  * @param chainOffset The offset of the chain entity in the class (-1 for non-entity classes).
  */
 extern "C" PLUGIN_API void SetEntData(int entityHandle, int offset, int64_t value, int size, bool changeState, int chainOffset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return;
@@ -332,7 +332,7 @@ extern "C" PLUGIN_API void SetEntData(int entityHandle, int offset, int64_t valu
  * @return The float value at the given memory location.
  */
 extern "C" PLUGIN_API double GetEntDataFloat(int entityHandle, int offset, int size) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot get '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return {};
@@ -352,7 +352,7 @@ extern "C" PLUGIN_API double GetEntDataFloat(int entityHandle, int offset, int s
  * @param chainOffset The offset of the chain entity in the class (-1 for non-entity classes).
  */
 extern "C" PLUGIN_API void SetEntDataFloat(int entityHandle, int offset, double value, int size, bool changeState, int chainOffset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return;
@@ -371,7 +371,7 @@ extern "C" PLUGIN_API void SetEntDataFloat(int entityHandle, int offset, double 
  * @return The string value at the given memory location.
  */
 extern "C" PLUGIN_API plg::string GetEntDataString(int entityHandle, int offset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot get '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return {};
@@ -390,7 +390,7 @@ extern "C" PLUGIN_API plg::string GetEntDataString(int entityHandle, int offset)
  * @param chainOffset The offset of the chain entity in the class (-1 for non-entity classes).
  */
 extern "C" PLUGIN_API void SetEntDataString(int entityHandle, int offset, const plg::string& value, bool changeState, int chainOffset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return;
@@ -409,7 +409,7 @@ extern "C" PLUGIN_API void SetEntDataString(int entityHandle, int offset, const 
  * @return The vector value at the given memory location.
  */
 extern "C" PLUGIN_API plg::vec3 GetEntDataVector(int entityHandle, int offset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot get '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return {};
@@ -428,7 +428,7 @@ extern "C" PLUGIN_API plg::vec3 GetEntDataVector(int entityHandle, int offset) {
  * @param chainOffset The offset of the chain entity in the class (-1 for non-entity classes).
  */
 extern "C" PLUGIN_API void SetEntDataVector(int entityHandle, int offset, const plg::vec3& value, bool changeState, int chainOffset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return;
@@ -449,7 +449,7 @@ extern "C" PLUGIN_API void SetEntDataVector(int entityHandle, int offset, const 
  * @return The entity handle at the given memory location.
  */
 extern "C" PLUGIN_API int GetEntDataEnt(int entityHandle, int offset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return INVALID_EHANDLE_INDEX;
@@ -468,7 +468,7 @@ extern "C" PLUGIN_API int GetEntDataEnt(int entityHandle, int offset) {
  * @param chainOffset The offset of the chain entity in the class (-1 for non-entity classes).
  */
 extern "C" PLUGIN_API void SetEntDataEnt(int entityHandle, int offset, int value, bool changeState, int chainOffset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return;
@@ -485,7 +485,7 @@ extern "C" PLUGIN_API void SetEntDataEnt(int entityHandle, int offset, int value
  * @param chainOffset The offset of the chain entity in the class (-1 for non-entity classes).
  */
 extern "C" PLUGIN_API void ChangeEntityState(int entityHandle, int offset, int chainOffset) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot change state of '{}' with invalid entity handle: {}\n", offset, entityHandle);
 		return;
@@ -1278,13 +1278,13 @@ extern "C" PLUGIN_API void SetEntSchemaEnt2(CEntityInstance* entity, const plg::
 
 	switch (schema::IsAtomicType(type, sizeof(CEntityHandle))) {
 		case schema::ElementType::Array:
-			reinterpret_cast<CEntityHandle*>(reinterpret_cast<intptr_t>(entity) + offset)[element] = CEntityHandle((uint32) value);
+			reinterpret_cast<CEntityHandle*>(reinterpret_cast<intptr_t>(entity) + offset)[element] = CEntityHandle(value);
 			break;
 		case schema::ElementType::Collection:
-			reinterpret_cast<CUtlVector<CEntityHandle>*>(reinterpret_cast<intptr_t>(entity) + offset)->Element(element) = CEntityHandle((uint32) value);
+			reinterpret_cast<CUtlVector<CEntityHandle>*>(reinterpret_cast<intptr_t>(entity) + offset)->Element(element) = CEntityHandle(value);
 			break;
 		case schema::ElementType::Single:
-			*reinterpret_cast<CEntityHandle*>(reinterpret_cast<intptr_t>(entity) + offset) = CEntityHandle((uint32) value);
+			*reinterpret_cast<CEntityHandle*>(reinterpret_cast<intptr_t>(entity) + offset) = CEntityHandle(value);
 			break;
 		default:
 			plg::print(LS_WARNING, "Schema field '{}::{}' is not a entity handle, but '{}'\n", className, memberName, type->m_sTypeName.Get());
@@ -1335,7 +1335,7 @@ extern "C" PLUGIN_API void NetworkStateChanged2(CEntityInstance* entity, const p
  * @return Size of array (in elements) or 0 if schema is not an array.
  */
 extern "C" PLUGIN_API int GetEntSchemaArraySize(int entityHandle, const plg::string& className, const plg::string& memberName) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1358,7 +1358,7 @@ extern "C" PLUGIN_API int GetEntSchemaArraySize(int entityHandle, const plg::str
  * @return An integer value at the given schema offset.
  */
 extern "C" PLUGIN_API int64_t GetEntSchema(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1381,7 +1381,7 @@ extern "C" PLUGIN_API int64_t GetEntSchema(int entityHandle, const plg::string& 
  * @return An integer value at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchema(int entityHandle, const plg::string& className, const plg::string& memberName, int64_t value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1404,7 +1404,7 @@ extern "C" PLUGIN_API void SetEntSchema(int entityHandle, const plg::string& cla
  * @return A float value at the given schema offset.
  */
 extern "C" PLUGIN_API double GetEntSchemaFloat(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1427,7 +1427,7 @@ extern "C" PLUGIN_API double GetEntSchemaFloat(int entityHandle, const plg::stri
  * @return An float value at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchemaFloat(int entityHandle, const plg::string& className, const plg::string& memberName, double value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1450,7 +1450,7 @@ extern "C" PLUGIN_API void SetEntSchemaFloat(int entityHandle, const plg::string
  * @return A string value at the given schema offset.
  */
 extern "C" PLUGIN_API plg::string GetEntSchemaString(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1473,7 +1473,7 @@ extern "C" PLUGIN_API plg::string GetEntSchemaString(int entityHandle, const plg
  * @return An string value at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchemaString(int entityHandle, const plg::string& className, const plg::string& memberName, const plg::string& value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1496,7 +1496,7 @@ extern "C" PLUGIN_API void SetEntSchemaString(int entityHandle, const plg::strin
  * @return A string value at the given schema offset.
  */
 extern "C" PLUGIN_API plg::vec3 GetEntSchemaVector3D(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1519,7 +1519,7 @@ extern "C" PLUGIN_API plg::vec3 GetEntSchemaVector3D(int entityHandle, const plg
  * @return A vector value at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchemaVector3D(int entityHandle, const plg::string& className, const plg::string& memberName, const plg::vec3& value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1542,7 +1542,7 @@ extern "C" PLUGIN_API void SetEntSchemaVector3D(int entityHandle, const plg::str
  * @return A string value at the given schema offset.
  */
 extern "C" PLUGIN_API plg::vec2 GetEntSchemaVector2D(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1565,7 +1565,7 @@ extern "C" PLUGIN_API plg::vec2 GetEntSchemaVector2D(int entityHandle, const plg
  * @return A vector value at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchemaVector2D(int entityHandle, const plg::string& className, const plg::string& memberName, const plg::vec2& value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1588,7 +1588,7 @@ extern "C" PLUGIN_API void SetEntSchemaVector2D(int entityHandle, const plg::str
  * @return A string value at the given schema offset.
  */
 extern "C" PLUGIN_API plg::vec4 GetEntSchemaVector4D(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return {};
@@ -1611,7 +1611,7 @@ extern "C" PLUGIN_API plg::vec4 GetEntSchemaVector4D(int entityHandle, const plg
  * @return A vector value at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchemaVector4D(int entityHandle, const plg::string& className, const plg::string& memberName, const plg::vec4& value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1636,7 +1636,7 @@ extern "C" PLUGIN_API void SetEntSchemaVector4D(int entityHandle, const plg::str
  * @return A string value at the given schema offset.
  */
 extern "C" PLUGIN_API int GetEntSchemaEnt(int entityHandle, const plg::string& className, const plg::string& memberName, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return INVALID_EHANDLE_INDEX;
@@ -1661,7 +1661,7 @@ extern "C" PLUGIN_API int GetEntSchemaEnt(int entityHandle, const plg::string& c
  * @return An entity handle at the given schema offset.
  */
 extern "C" PLUGIN_API void SetEntSchemaEnt(int entityHandle, const plg::string& className, const plg::string& memberName, int value, bool changeState, int element) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot set '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
@@ -1685,7 +1685,7 @@ extern "C" PLUGIN_API void SetEntSchemaEnt(int entityHandle, const plg::string& 
  * @param memberName The name of the member to be set.
  */
 extern "C" PLUGIN_API void NetworkStateChanged(int entityHandle, const plg::string& className, const plg::string& memberName) {
-	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle((uint32) entityHandle));
+	CEntityInstance* entity = g_pGameEntitySystem->GetEntityInstance(CEntityHandle(entityHandle));
 	if (!entity) {
 		plg::print(LS_WARNING, "Cannot change state of '{}::{}' with invalid entity handle: {}\n", className, memberName, entityHandle);
 		return;
